requirements.txt
# ── Core ────────────────────────────────────────────────
streamlit>=1.35.0

# ── LangChain ecosystem ──────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-pinecone>=0.1.1
langchain-groq>=0.1.4

# ── Vector DB ────────────────────────────────────────────
pinecone-client>=3.2.0

# ── Embeddings ───────────────────────────────────────────
sentence-transformers>=2.7.0
torch>=2.1.0                  # CPU version is fine
transformers>=4.40.0

# ── Document loading ─────────────────────────────────────
pypdf>=4.2.0                  # PDF loading
unstructured>=0.13.0          # optional, improves PDF parsing

# ── Utilities ────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
